Move spinner construction into own module in hopes of making it more obvious where the spinners are created.
